Complete exercises 4.20-4.22 (extract middleware, authenticate DELETE)

+++ part4/bloglist/controllers/blogs.js
@@ -1,15 +1,6 @@
-const jwt = require('jsonwebtoken');
 const blogsRouter = require('express').Router();
 const Blog = require('../models/blog');
 const User = require('../models/user');
-
-const getTokenFrom = (request) => {
-  const authorization = request.get('authorization');
-  if (authorization && authorization.toLowerCase().startsWith('bearer ')) {
-    return authorization.substring(7);
-  }
-  return null;
-};
 
 blogsRouter.get('/', async (request, response) => {
   const blogs = await Blog
@@ -38,14 +29,7 @@ blogsRouter.post('/', async (request, response) => {
     likes,
   } = request.body;
 
-  const token = getTokenFrom(request);
-  const decodedToken = jwt.verify(token, process.env.SECRET);
-
-  if (!decodedToken.id) {
-    return response.status(401).json({ error: 'token missing or invalid' });
-  }
-
-  const user = await User.findById(decodedToken.id);
+  const user = await User.findById(request.user.id);
 
   const blog = new Blog({
     title,
@@ -64,8 +48,14 @@ blogsRouter.post('/', async (request, response) => {
 /* eslint-enable no-underscore-dangle */
 
 blogsRouter.delete('/:id', async (request, response) => {
+  const blog = await Blog.findById(request.params.id);
+
+  if (blog.user.toString() !== request.user.id.toString()) {
+    return response.status(401).json({ error: 'cannot delete another user\'s blog' });
+  }
+
   await Blog.findByIdAndRemove(request.params.id);
-  response.status(204).end();
+  return response.status(204).end();
 });
 
 blogsRouter.put('/:id', async (request, response) => {
